+++ apps/client/tailwind.config.js
@@ -1,5 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 export default {
+  darkMode: 'class',
   content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}', './node_modules/@heroui/react/dist/**/*.{js,ts,jsx,tsx}'],
   theme: {
     extend: {